matrix: replace uint64 bound to int
Having int as bounds simplifies the for loops setup, as well as enables
the tests for negative bounds with a controlled informative error in the
affirmative case, instead of having an overflow uint64 error which can't
be modified.

Removed commented code.

+++ matrix/matrix.go
@@ -8,21 +8,22 @@ import (
 
 // Interface that condenses basic linear algebra operations operations
 type NumberArray interface {
-	GetColumns() uint64
-	GetRows() uint64
+	GetColumns() int
+	GetRows() int
 	Transpose()
-	SetValue(i, j uint64, v float64) error
-	GetValue(i, j uint64) (float64, error)
+	SetValue(i, j int, v float64) error
+	GetValue(i, j int) (float64, error)
 }
 
 // Basic type used for all linear algebra operations
 type matrix struct {
 	matrix [][]float64
-	rows   uint64
-	cols   uint64
+	rows   int
+	cols   int
 }
 
-func checkPositiveBounds(rows, cols uint64) (b bool, err error) {
+// Checks that the sizes of the given matrix are non-zero and positive
+func checkPositiveBounds(rows, cols int) (b bool, err error) {
 	if rows < 1 {
 		return b, fmt.Errorf("Can't create a matrix with %v rows", rows)
 	}
@@ -33,13 +34,13 @@ func checkPositiveBounds(rows, cols uint64) (b bool, err error) {
 }
 
 // Creates a matrix of the given rows and colums
-func NewMatrix(rows, cols uint64) (m *matrix, err error) {
+func NewMatrix(rows, cols int) (m *matrix, err error) {
 	if ok, err := checkPositiveBounds(rows, cols); !ok {
 		return m, err
 	}
 	m = new(matrix)
 	m.matrix = make([][]float64, rows)
-	for i := uint64(0); i < rows; i++ {
+	for i := 0; i < rows; i++ {
 		m.matrix[i] = make([]float64, cols)
 	}
 	m.rows = rows
@@ -48,27 +49,22 @@ func NewMatrix(rows, cols uint64) (m *matrix, err error) {
 }
 
 // Creates a new ColumnVector which is nothing more than a one-column matrix
-func NewColumnVector(rows uint64) (*matrix, error) {
+func NewColumnVector(rows int) (*matrix, error) {
 	return NewMatrix(rows, 1)
 }
 
-// // Gets the dimensions of the matrix as rows, columns pair
-// func (m *matrix) GetDimensions() (uint64, uint64) {
-// 	return m.rows, m.cols
-// }
-
 // Gets the number of rows of the matrix
-func (m *matrix) GetRows() uint64 {
+func (m *matrix) GetRows() int {
 	return m.rows
 }
 
 // Gets the number of columns of the matrix
-func (m *matrix) GetColumns() uint64 {
+func (m *matrix) GetColumns() int {
 	return m.cols
 }
 
 // Gets the value in index i, j, or exits program if error
-func (m *matrix) GetValue(i, j uint64) (f float64, err error) {
+func (m *matrix) GetValue(i, j int) (f float64, err error) {
 	if ok, err := m.checkBounds(i, j); !ok {
 		return f, err
 	}
@@ -76,7 +72,7 @@ func (m *matrix) GetValue(i, j uint64) (f float64, err error) {
 }
 
 // Sets a value v in the row i, and column j. i.e: m[i][j] = v
-func (m *matrix) SetValue(i, j uint64, v float64) (err error) {
+func (m *matrix) SetValue(i, j int, v float64) (err error) {
 	if ok, err := m.checkBounds(i, j); !ok {
 		return err
 	}
@@ -85,7 +81,7 @@ func (m *matrix) SetValue(i, j uint64, v float64) (err error) {
 }
 
 // Checks that the indexes i and j are withing bounds
-func (a *matrix) checkBounds(i, j uint64) (b bool, err error) {
+func (a *matrix) checkBounds(i, j int) (b bool, err error) {
 	if i < 0 || i >= a.rows {
 		return b, fmt.Errorf("Index %v is out of bounds for rows with size %v", i, a.rows)
 
@@ -164,10 +160,10 @@ func Dot(a, b NumberArray) (resultingMatrix NumberArray, err error) {
 	if err != nil {
 		return resultingMatrix, err
 	}
-	for i := uint64(0); i < a.GetRows(); i++ {
-		for j := uint64(0); j < b.GetColumns(); j++ {
+	for i := 0; i < a.GetRows(); i++ {
+		for j := 0; j < b.GetColumns(); j++ {
 			sum := 0.0
-			for k := uint64(0); k < a.GetColumns(); k++ {
+			for k := 0; k < a.GetColumns(); k++ {
 				operandA, _ := a.GetValue(i, k)
 				operandB, _ := b.GetValue(k, j)
 				sum += operandA * operandB
@@ -177,23 +173,3 @@ func Dot(a, b NumberArray) (resultingMatrix NumberArray, err error) {
 	}
 	return resultingMatrix, err
 }
-
-//
-// func matrixMultiplication(a, b NumberArray) (resultingMatrix NumberArray, err error) {
-// 	if a.GetColumns() != b.GetRows() {
-// 		fmt.Errorf("Can't multiply matrices that don't satisfy multiplication criteria")
-// 		err = errors.New("Can't multiply matrices that don't satisfy multiplication criteria")
-// 		return nil, err
-// 	}
-// 	resultingMatrix = newMatrix(a.GetRows(), b.GetColumns())
-// 	for i := uint64(0); i < a.GetRows(); i++ {
-// 		for j := uint64(0); j < b.GetColumns(); j++ {
-// 			sum := 0.0
-// 			for k := uint64(0); k < a.GetColumns(); k++ {
-// 				sum += a.GetValue(i, k) * b.GetValue(k, j)
-// 			}
-// 			resultingMatrix.SetValue(i, j, sum)
-// 		}
-// 	}
-// 	return resultingMatrix, nil
-// }
